requirements.txt
pytest
pyyaml
requests
websocket-client


